Migrate BigFretBoard to typescript
