+++ package.json
@@ -40,6 +40,7 @@
     ]
   },
   "devDependencies": {
+    "@types/lodash": "^4.14.171",
     "@types/react-router-dom": "^5.1.8",
     "@types/styled-components": "^5.1.11"
   }
